see changelog 1.0.3

+++ api/adm_deleteEntryById.php
@@ -37,8 +37,13 @@
                         echo "success";
                     }                   
                 } elseif($_GET["table"] == "users") {
-                    $object->delete();
-                    echo "success";
+                    if($object->firstname == "Administrator") {
+                        echo "ERROR: cannot delete builtin admin User.";
+                    } 
+                    else {
+                        $object->delete();
+                        echo "success";
+                    }                    
                 }                
             }
             else {
